Extract `rescue-theme-error`
