+++ app/assets/javascripts/discourse/lib/utilities.js.es6
@@ -1,5 +1,4 @@
 import { escape } from "pretty-text/sanitizer";
-import { ajax } from "discourse/lib/ajax";
 import toMarkdown from "discourse/lib/to-markdown";
 
 const homepageSelector = "meta[name=discourse_current_homepage]";
@@ -382,40 +381,6 @@ export function postRNWebviewMessage(prop, value) {
   if (window.ReactNativeWebView !== undefined) {
     window.ReactNativeWebView.postMessage(JSON.stringify({ [prop]: value }));
   }
-}
-
-function reportToLogster(name, error) {
-  const data = {
-    message: `${name} theme/component is throwing errors`,
-    stacktrace: error.stack
-  };
-
-  ajax("/logs/report_js_error", {
-    data,
-    type: "POST",
-    cache: false
-  });
-}
-// this function is used in lib/theme_javascript_compiler.rb
-export function rescueThemeError(name, error, api) {
-  /* eslint-disable-next-line no-console */
-  console.error(`"${name}" error:`, error);
-  reportToLogster(name, error);
-
-  const currentUser = api.getCurrentUser();
-  if (!currentUser || !currentUser.admin) {
-    return;
-  }
-
-  const path = `${Discourse.BaseUri}/admin/customize/themes`;
-  const message = I18n.t("themes.broken_theme_alert", {
-    theme: name,
-    path: `<a href="${path}">${path}</a>`
-  });
-  const alertDiv = document.createElement("div");
-  alertDiv.classList.add("broken-theme-alert");
-  alertDiv.innerHTML = `⚠️ ${message}`;
-  document.body.prepend(alertDiv);
 }
 
 const CODE_BLOCKS_REGEX = /^(    |\t).*|`[^`]+`|^```[^]*?^```|\[code\][^]*?\[\/code\]/gm;
